Remove test-primitives
