--- test/primitives/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "test-primitives"
-version = "0.1.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2018"
-
-[dependencies]
-# Substrate dependencies
-sp-core = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "rococo-branch" }
-sp-runtime = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "rococo-branch" }
-
-[features]
-default = [ "std" ]
-std = [
-	"sp-runtime/std",
-	"sp-core/std"
-]
